Refactor and add more error handling
this commit does quite a bit of refactoring of the consolidated logic
into their dedicated classes. Now there's a Driver class to actually
parse the command line input and that delegates work to the Validator,
BundleGenerator and so on. The Validator now does all the config file
validation and either successfully executes or sys exits