Generate or use existing snapshot file
the logic for this is pretty simple but maybe prone to some mistakes.
Each bundle checks in it's path for a file matching a simple regex of
the default_locale (the expecation is that only one file will match the
regex _{default_locale}) without any fallback mechanism. A snapshot file
with the {default_locale path}.snapshot either exists at the bundle's
path or it doesn't. If it doesn't, it's created and it'll be a duplicate
of the default locale file in the path