Generate list of potential candidate for translations
the logic is simple for now. Take a "subtraction" of the keys in the
default_locale file and the snapshot. Every key that differs is a
potential addition. Get the corresponding message associated with each
of these "new" keys and put them in a list. There's no way to currently
handle if an existing key was renamed or updated. They will all show up
as "new" conceptually. When do we edit the snapshot to rebalance this
change?