Build json of additions and missing values
iterate through each bunlde and within each bundle compare the default
locale's values with the snapshot to determine which new strings need to
be translated. Compare the snapshot keys with keys in every other locale
to determine which locales are missing keys and collect the value of
those keys from the snapshot file. They need to be sent out for
translation again